refactor index page for schema reg
